+++ src/test/java/br/com/fiap/entities/Paciente.java
@@ -4,7 +4,7 @@ import java.time.LocalDate;
 
 public class Paciente {
 
-    private int id;
+    private Long id;
     private String nome;
     private String cpf;
     private String telefone;
@@ -16,7 +16,7 @@ public class Paciente {
     public Paciente() {
     }
 
-    public Paciente(int id, String nome, String cpf, String telefone, String email, LocalDate dataNascimento, boolean ativo, String planoOdontologico) {
+    public Paciente(Long id, String nome, String cpf, String telefone, String email, LocalDate dataNascimento, boolean ativo, String planoOdontologico) {
         this.id = id;
         this.nome = nome;
         this.cpf = cpf;
@@ -27,13 +27,9 @@ public class Paciente {
         this.planoOdontologico = planoOdontologico;
     }
 
-    public int getId() {
-        return id;
-    }
+    public Long getId() { return id; }
 
-    public void setId(int id) {
-        this.id = id;
-    }
+    public void setId(Long id) { this.id = id; }
 
     public String getNome() {
         return nome;
